cleaned the code
