+++ public/menu1.json
@@ -1,5 +1,5 @@
 {
-    "random-dishes": [
+    "dishes": [
         {
             "name": "Roast Duck Breast",
             "recipe": "Roasted duck breast (served pink) with gratin potato and a griottine cherry sauce",
