--- src/components/Home/Banner/Banner.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import '../css/Home.css'
-const Banner = () => {
-    return (
-        <div className="text-center text-white flex relative top-0 justify-center items-center min-h-[700px] bg-image">
-           <div className='text-style'> 
-           <h1 className=' uppercase text-7xl '>Culinary Cloud</h1>
-           <small className='text-lg'>123 ABC Street, Uni 21, Bangladesh</small>
-           </div>
-        </div>
-            
-
-    );
-};
-
-export default Banner;
