+++ src/components/Home/Menu.jsx
@@ -1,22 +1,21 @@
 import React, { useState, useEffect } from "react";
-import {Card} from '../../shared/Card/Card'
-import Title from "../../shared/Title/Title";
-const Menu = () => {
-  const [menuItems, setMenuItems] = useState([]);
+import useItems from "../../Hooks/useItems";
+import {Card} from '../shared/Card/Card'
+import Title from '../shared/Title/Title';
 
-  useEffect(() => {
-    fetch("menu.json")
-      .then((res) => res.json())
-      .then((data) => setMenuItems(data));
-  }, []);
-  console.log(menuItems)
+const Menu = () => {
+  const [items,loading] = useItems();
+    if(loading){
+        return <div>loading...</div>
+    }
+  
   return (
     <div >
     
       <Title type={{ smallHeading: 'Check it out', title: 'FROM OUR MENU', border: 'black' }}></Title>
       <div class="grid grid-cols-2 gap-5 px-10">
         
-          {menuItems.map((menuItem) => (
+          {items.dishes.map((menuItem) => (
              <div className="flex space-x-4">
              <Card>
                  {menuItem}
