+++ src/components/Menu/Soup.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Card } from '../shared/Card/Card';
-import '../shared/css/Common.css'   /*bg-fixed*/
-
+import './Menu.css'
 export const Soup = ({ items }) => {
     console.log(items)
