--- src/components/Shop/Banner/Banner.css
@@ -1,6 +1,0 @@
-.bg-image2{
-    background-image: url('../../../assets/asset/home/banner2.jpg');
-    background-size: cover;
-    height: 100vh;
-    width: 100%;
-}
