+++ src/components/Shop/Drinks.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import useItems from '../../../Hooks/useItems';
-import Cards from '../Cards/Cards';
+import useItems from '../../Hooks/useItems';
+import Cards from '../shared/Card/Card2';
 
-const Pizza = () => {
-    const [items,,loading] = useItems();
+const Drinks = () => {
+    const [items,loading] = useItems();
     if(loading){
         return <div>loading...</div>
     }
@@ -11,7 +11,7 @@ const Pizza = () => {
         <div>
             <div className="grid grid-cols-3 gap-x-2 gap-y-10 mt-10 place-items-center">
             {
-                    items?.pizza.map((item) => 
+                    items?.offer?.map((item) => 
                     <Cards>
                         {item}
                     </Cards>
@@ -23,4 +23,4 @@ const Pizza = () => {
     );
 };
 
-export default Pizza;
+export default Drinks;
