+++ src/components/Shop/Pizza.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
-import useItems from '../../../Hooks/useItems';
-import Cards from '../Cards/Cards';
-const Soups = () => {
-    const [items,,loading] = useItems();
+import useItems from '../../Hooks/useItems';
+import Cards from '../shared/Card/Card2';
+
+const Pizza = () => {
+    const [items,loading] = useItems();
     if(loading){
         return <div>loading...</div>
     }
-    console.log(items)
     return (
         <div>
             <div className="grid grid-cols-3 gap-x-2 gap-y-10 mt-10 place-items-center">
             {
-                    items?.offer?.map((item) => 
+                    items?.pizza.map((item) => 
                     <Cards>
                         {item}
                     </Cards>
@@ -23,4 +23,4 @@ const Soups = () => {
     );
 };
 
-export default Soups;
+export default Pizza;
