+++ src/components/Shop/Soups.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import useItems from '../../../Hooks/useItems';
-import Cards from '../Cards/Cards';
-
-const Drinks = () => {
-    const [items,,loading] = useItems();
+import useItems from '../../Hooks/useItems';
+import Cards from '../shared/Card/Card2';
+const Soups = () => {
+    const [items,loading] = useItems();
     if(loading){
         return <div>loading...</div>
     }
+    console.log(items)
     return (
         <div>
             <div className="grid grid-cols-3 gap-x-2 gap-y-10 mt-10 place-items-center">
@@ -23,4 +23,4 @@ const Drinks = () => {
     );
 };
 
-export default Drinks;
+export default Soups;
