+++ src/components/shared/Banner2/Banner2.css
@@ -1,0 +1,20 @@
+.bg-img{
+    height: 100vh;
+    width: 100%;
+  
+}
+
+.section{
+    background: linear-gradient(to right, rgba(0, 0, 0, 0.50) 0%,rgba(0, 0, 0, 0.50) 70%);
+    opacity: 0.8;
+    height: 80vh;
+    width: 90%;
+}
+
+.bg-menu-banner{
+    background-image: url('../../../assets/asset/menu/banner3.jpg');
+}
+
+.bg-shop-banner{
+    background-image: url('../../../assets/asset/home/banner.jpg');
+}
