+++ src/components/shared/Card/Card2.jsx
