--- src/components/shared/css/Common.css
@@ -1,5 +1,0 @@
-.bg-fixed{
-    background-size: cover;
-    background-attachment: fixed;
-  
-}
