--- src/pages/Menu/Menu.css
@@ -1,3 +1,0 @@
-.menu-bg{
-    background-image: url('../../assets/others/menu-bg.png');
-}
