+++ src/pages/Menu/Menu.jsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import { Banner } from '../../components/Menu/Banner';
 import { Dessert } from '../../components/Menu/Dessert';
 import { Offer } from '../../components/Menu/Offer';
 import { Pizza } from '../../components/Menu/Pizza';
 import { Salad } from '../../components/Menu/Salad';
 import { Soup } from '../../components/Menu/Soup';
-import './Menu.css'
+import Banner2 from '../../components/shared/Banner2/Banner2';
+
 
 const Menu = () => {
 
@@ -19,7 +19,7 @@ const Menu = () => {
 
     return (
         <div>
-            <Banner></Banner>
+            <Banner2 banner = {{title: 'Our Menu', description:'Would you like to try a dish?', image:'bg-menu-banner'}}></Banner2>
             <div>
                 {items && <Offer items={items.offer}></Offer>}
                 {items && <Dessert items={items.dessert}></Dessert>}
